requirements.txt
altair==5.5.0
altgraph==0.17.4
asyncio==3.4.3
attrs==24.3.0
audiolazy==0.6
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
clipboard==0.0.4
colormap==1.1.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
MIDIUtil==1.2.1
mido==1.3.3
narwhals==1.22.0
Nuitka==2.6.4
numpy==2.1.1
opencv-python-headless==4.11.0.86
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pretty_midi==0.2.10
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyparsing==3.1.4
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.5.0
six==1.16.0
smmap==5.0.2
st-copy-to-clipboard==0.1.6
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
zstandard==0.23.0
